+++ Ray.hpp
@@ -2,6 +2,7 @@
 
 #include <glm/vec3.hpp>
 #include <glm/vec4.hpp>
+#include "Vec3.hpp"
 
 class ray
 {
@@ -21,7 +22,7 @@ public:
         : origin(neworigin), dir(newdir) {}
 
     // Member Functions
-    glm::vec3 at(float t) const
+    glm::vec3 at(double t) const
     {
         return origin + t * dir;
     }
